refactor: extract FormRechercheOffres
