+++ pages/recherche-offres.tsx
@@ -1,16 +1,9 @@
 import { withTransaction } from '@elastic/apm-rum-react'
 import { GetServerSideProps } from 'next'
-import React, { FormEvent, useEffect, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 
-import RadioButton from 'components/action/RadioButton'
-import RechercheOffresEmploiMain from 'components/offres/RechercheOffresEmploiMain'
-import RechercheOffresEmploiSecondary from 'components/offres/RechercheOffresEmploiSecondary'
-import RechercheServicesCiviquesMain from 'components/offres/RechercheServicesCiviquesMain'
-import RechercheServicesCiviquesSecondary from 'components/offres/RechercheServicesCiviquesSecondary'
+import FormRechercheOffres from 'components/offres/FormRechercheOffres'
 import ResultatsRechercheOffre from 'components/offres/ResultatsRechercheOffres'
-import Button from 'components/ui/Button/Button'
-import { Etape } from 'components/ui/Form/Etape'
-import IconComponent, { IconName } from 'components/ui/IconComponent'
 import FailureAlert from 'components/ui/Notifications/FailureAlert'
 import {
   BaseOffre,
@@ -38,7 +31,6 @@ type WithHasError<T> = T & { hasError: boolean }
 type RechercheOffresProps = PageProps & {
   partageSuccess?: boolean
 }
-
 function RechercheOffres({ partageSuccess }: RechercheOffresProps) {
   const referentielService =
     useDependance<ReferentielService>('referentielService')
@@ -50,13 +42,10 @@ function RechercheOffres({ partageSuccess }: RechercheOffresProps) {
   )
 
   const [typeOffre, setTypeOffre] = useState<TypeOffre | undefined>()
-  const [showMoreFilters, setShowMoreFilters] = useState<boolean>(false)
-
-  const [countCriteres, setCountCriteres] = useState<number>(0)
   const [queryOffresEmploi, setQueryOffresEmploi] = useState<
     WithHasError<SearchOffresEmploiQuery>
   >({ hasError: false })
-  const [queryServiceCivique, setQueryServiceCivique] = useState<
+  const [queryServicesCiviques, setQueryServicesCiviques] = useState<
     WithHasError<SearchServicesCiviquesQuery>
   >({ hasError: false })
 
@@ -72,16 +61,8 @@ function RechercheOffres({ partageSuccess }: RechercheOffresProps) {
   if (partageSuccess) initialTracking += ' - Partage offre succès'
   const [trackingTitle, setTrackingTitle] = useState<string>(initialTracking)
 
-  const formIsInvalid =
-    queryOffresEmploi.hasError || queryServiceCivique.hasError
-
-  async function rechercherPremierePage(e: FormEvent<HTMLFormElement>) {
-    e.preventDefault()
-    if (isSearching) return
-    if (!typeOffre) return
-    if (formIsInvalid) return
+  async function rechercherPremierePage() {
     nettoyerResultats()
-
     rechercherOffres({ page: 1 })
   }
 
@@ -130,7 +111,7 @@ function RechercheOffres({ partageSuccess }: RechercheOffresProps) {
     offres: BaseServiceCivique[]
     metadonnees: MetadonneesOffres
   }> {
-    const { hasError, ...query } = queryServiceCivique
+    const { hasError, ...query } = queryServicesCiviques
     return servicesCiviquesService.searchServicesCiviques(query, page)
   }
 
@@ -143,13 +124,8 @@ function RechercheOffres({ partageSuccess }: RechercheOffresProps) {
   }
 
   useEffect(() => {
-    setQueryOffresEmploi({ hasError: false })
-    setQueryServiceCivique({ hasError: false })
-  }, [typeOffre])
-
-  useEffect(() => {
     nettoyerResultats()
-  }, [queryOffresEmploi, queryServiceCivique])
+  }, [queryOffresEmploi, queryServicesCiviques])
 
   useMatomo(trackingTitle)
 
@@ -162,80 +138,21 @@ function RechercheOffres({ partageSuccess }: RechercheOffresProps) {
         />
       )}
 
-      <form
-        onSubmit={rechercherPremierePage}
-        className={
-          offres !== undefined || isSearching
-            ? 'bg-primary_lighten p-6 mb-10 rounded-small'
-            : ''
-        }
-      >
-        <Etape numero={1} titre='Sélectionner un type d’offre'>
-          <div className='flex flex-wrap'>
-            <RadioButton
-              isSelected={typeOffre === TypeOffre.EMPLOI}
-              onChange={() => setTypeOffre(TypeOffre.EMPLOI)}
-              name='type-offre'
-              id='type-offre--emploi'
-              label='Offre d’emploi'
-            />
-            <RadioButton
-              isSelected={typeOffre === TypeOffre.SERVICE_CIVIQUE}
-              onChange={() => setTypeOffre(TypeOffre.SERVICE_CIVIQUE)}
-              name='type-offre'
-              id='type-offre--service-civique'
-              label='Service civique'
-            />
-          </div>
-        </Etape>
-
-        {getRechercheMain()}
-
-        {typeOffre && (
-          <div className='flex justify-end mb-6'>
-            <button
-              type='button'
-              onClick={() => setShowMoreFilters(!showMoreFilters)}
-              className='mr-12'
-            >
-              Voir {showMoreFilters ? 'moins' : 'plus'} de critères
-              <IconComponent
-                name={
-                  showMoreFilters ? IconName.ChevronUp : IconName.ChevronDown
-                }
-                className='h-4 w-4 fill-primary inline ml-2'
-                aria-hidden={true}
-                focusable={false}
-              ></IconComponent>
-            </button>
-          </div>
+      <FormRechercheOffres
+        isSearching={isSearching}
+        hasResults={offres !== undefined}
+        fetchCommunes={referentielService.getCommunes.bind(referentielService)}
+        fetchCommunesEtDepartements={referentielService.getCommunesEtDepartements.bind(
+          referentielService
         )}
-
-        {showMoreFilters && getRechercheSecondary()}
-
-        {typeOffre && (
-          <>
-            <div className='mt-5 mb-4 text-center'>
-              [{countCriteres}] critère{countCriteres > 1 && 's'} sélectionné
-              {countCriteres > 1 && 's'}
-            </div>
-
-            <Button
-              type='submit'
-              className='mx-auto'
-              disabled={formIsInvalid || isSearching}
-            >
-              <IconComponent
-                name={IconName.Search}
-                focusable={false}
-                aria-hidden={true}
-                className='w-4 h-4 mr-2'
-              />
-              Rechercher
-            </Button>
-          </>
-        )}
-      </form>
+        stateTypeOffre={[typeOffre, setTypeOffre]}
+        stateQueryOffresEmploi={[queryOffresEmploi, setQueryOffresEmploi]}
+        stateQueryServicesCiviques={[
+          queryServicesCiviques,
+          setQueryServicesCiviques,
+        ]}
+        onNouvelleRecherche={rechercherPremierePage}
+      />
 
       <ResultatsRechercheOffre
         isSearching={isSearching}
@@ -247,56 +164,6 @@ function RechercheOffres({ partageSuccess }: RechercheOffresProps) {
       />
     </>
   )
-
-  function getRechercheMain(): JSX.Element | null {
-    switch (typeOffre) {
-      case TypeOffre.EMPLOI:
-        return (
-          <RechercheOffresEmploiMain
-            fetchCommunesEtDepartements={referentielService.getCommunesEtDepartements.bind(
-              referentielService
-            )}
-            query={queryOffresEmploi}
-            onQueryUpdate={setQueryOffresEmploi}
-          />
-        )
-      case TypeOffre.SERVICE_CIVIQUE:
-        return (
-          <RechercheServicesCiviquesMain
-            fetchCommunes={referentielService.getCommunes.bind(
-              referentielService
-            )}
-            query={queryServiceCivique}
-            onQueryUpdate={setQueryServiceCivique}
-          />
-        )
-      default:
-        return null
-    }
-  }
-
-  function getRechercheSecondary(): JSX.Element | null {
-    switch (typeOffre) {
-      case TypeOffre.EMPLOI:
-        return (
-          <RechercheOffresEmploiSecondary
-            onCriteresChange={setCountCriteres}
-            query={queryOffresEmploi}
-            onQueryUpdate={setQueryOffresEmploi}
-          />
-        )
-      case TypeOffre.SERVICE_CIVIQUE:
-        return (
-          <RechercheServicesCiviquesSecondary
-            onCriteresChange={setCountCriteres}
-            query={queryServiceCivique}
-            onQueryUpdate={setQueryServiceCivique}
-          />
-        )
-      default:
-        return null
-    }
-  }
 }
 
 export const getServerSideProps: GetServerSideProps<
